REFACTOR: create sub packages
